Closes #61, adds "recent high fire season" layer.
This is done by applying a different style to the same shapefile used for historical data, so there are some changes to pass that WMS request (styles) across.  Also, since 'name' was used as a unique key field in a few places, the code now checks (and uses) "layerName" for the name of the layer, which need not be unique.